Convert HeaderBlock into RSC
