+++ apps/store/src/components/Header/ShoppingCartMenuItem.tsx
@@ -1,3 +1,4 @@
+'use client'
 import styled from '@emotion/styled'
 import Link from 'next/link'
 import { theme } from 'ui'
