+++ apps/store/src/components/LayoutWithMenu/LayoutWithMenu.tsx
@@ -1,7 +1,7 @@
 import { clsx } from 'clsx'
-import { Fragment, type ReactElement } from 'react'
+import { type ReactElement } from 'react'
 import { FooterBlock } from '@/blocks/FooterBlock'
-import { HeaderBlock } from '@/blocks/HeaderBlock'
+import { HeaderBlock } from '@/blocks/HeaderBlock/HeaderBlock'
 import { wrapper } from '@/components/LayoutWithMenu/LayoutWithMenu.css'
 import type { GlobalStory } from '@/services/storyblok/storyblok'
 import { filterByBlockType } from '@/services/storyblok/Storyblok.helpers'
